ccl: refactor tenant status server test to statusccl package
Previously, tenant status server tests live within serverccl package.
However, this caused all nightly test failures to automaticallly
notify server team through GitHub code owner file.
This commit moves all tenant status server into statusccl package
and update statusccl package code owner to SQL O11y.

Release note: None
