+++ pkg/BUILD.bazel
@@ -30,6 +30,7 @@ ALL_TESTS = [
     "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver_test",
     "//pkg/ccl/oidcccl:oidcccl_test",
     "//pkg/ccl/partitionccl:partitionccl_test",
+    "//pkg/ccl/serverccl/statusccl:statusccl_test",
     "//pkg/ccl/serverccl:serverccl_test",
     "//pkg/ccl/sqlproxyccl/denylist:denylist_test",
     "//pkg/ccl/sqlproxyccl/idle:idle_test",
