+++ pkg/ccl/serverccl/BUILD.bazel
@@ -2,27 +2,9 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "serverccl",
-    srcs = [
-        "doc.go",
-        "tenant_test_utils.go",
-    ],
+    srcs = ["doc.go"],
     importpath = "github.com/cockroachdb/cockroach/pkg/ccl/serverccl",
     visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/base",
-        "//pkg/roachpb:with-mocks",
-        "//pkg/security",
-        "//pkg/server/serverpb",
-        "//pkg/sql/pgwire",
-        "//pkg/sql/sqlstats/persistedsqlstats",
-        "//pkg/sql/tests",
-        "//pkg/testutils/serverutils",
-        "//pkg/testutils/sqlutils",
-        "//pkg/util/httputil",
-        "//pkg/util/log",
-        "//pkg/util/protoutil",
-        "@com_github_stretchr_testify//require",
-    ],
 )
 
 go_test(
@@ -33,8 +15,6 @@ go_test(
         "main_test.go",
         "role_authentication_test.go",
         "server_sql_test.go",
-        "tenant_grpc_test.go",
-        "tenant_status_test.go",
         "tenant_vars_test.go",
     ],
     embed = [":serverccl"],
@@ -45,20 +25,14 @@ go_test(
         "//pkg/ccl/utilccl",
         "//pkg/ccl/utilccl/licenseccl",
         "//pkg/roachpb:with-mocks",
-        "//pkg/rpc",
         "//pkg/security",
         "//pkg/security/securitytest",
         "//pkg/server",
         "//pkg/server/serverpb",
         "//pkg/sql",
-        "//pkg/sql/catalog/catconstants",
-        "//pkg/sql/catalog/descpb",
-        "//pkg/sql/idxusage",
         "//pkg/sql/pgwire/pgcode",
-        "//pkg/sql/sqlstats",
         "//pkg/sql/tests",
         "//pkg/testutils/serverutils",
-        "//pkg/testutils/skip",
         "//pkg/testutils/sqlutils",
         "//pkg/testutils/testcluster",
         "//pkg/util",
