+++ pkg/ccl/serverccl/statusccl/BUILD.bazel
@@ -1,0 +1,59 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "statusccl",
+    srcs = ["tenant_test_utils.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/serverccl/statusccl",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/base",
+        "//pkg/roachpb:with-mocks",
+        "//pkg/security",
+        "//pkg/server/serverpb",
+        "//pkg/sql/pgwire",
+        "//pkg/sql/sqlstats/persistedsqlstats",
+        "//pkg/sql/tests",
+        "//pkg/testutils/serverutils",
+        "//pkg/testutils/sqlutils",
+        "//pkg/util/httputil",
+        "//pkg/util/log",
+        "//pkg/util/protoutil",
+        "@com_github_stretchr_testify//require",
+    ],
+)
+
+go_test(
+    name = "statusccl_test",
+    srcs = [
+        "main_test.go",
+        "tenant_grpc_test.go",
+        "tenant_status_test.go",
+    ],
+    embed = [":statusccl"],
+    deps = [
+        "//pkg/base",
+        "//pkg/ccl",
+        "//pkg/ccl/kvccl",
+        "//pkg/ccl/utilccl",
+        "//pkg/roachpb:with-mocks",
+        "//pkg/rpc",
+        "//pkg/security",
+        "//pkg/security/securitytest",
+        "//pkg/server",
+        "//pkg/server/serverpb",
+        "//pkg/sql/catalog/catconstants",
+        "//pkg/sql/catalog/descpb",
+        "//pkg/sql/idxusage",
+        "//pkg/sql/sqlstats",
+        "//pkg/sql/tests",
+        "//pkg/testutils/serverutils",
+        "//pkg/testutils/skip",
+        "//pkg/testutils/sqlutils",
+        "//pkg/testutils/testcluster",
+        "//pkg/util/leaktest",
+        "//pkg/util/log",
+        "//pkg/util/randutil",
+        "//pkg/util/timeutil",
+        "@com_github_stretchr_testify//require",
+    ],
+)
