+++ pkg/ccl/serverccl/statusccl/tenant_status_test.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package serverccl
+package statusccl
 
 import (
 	"context"
