+++ pkg/ccl/serverccl/statusccl/tenant_test_utils.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package serverccl
+package statusccl
 
 import (
 	"context"
